extract mappers to their own files
