+++ src/store/actions.js
@@ -2,7 +2,9 @@
 
 // $FlowFixMe
 import api from '@molgenis/molgenis-api-client'
-import * as mappers from '../utils/mappers'
+import activeFiltersToRsqlMapper from '../utils/active-filters-to-rsql-mapper'
+import aggregateDataToChartDataMapper from '../utils/aggregate-data-to-chart-data-mapper'
+import subjectMetadataToFilterMapper from '../utils/subject-metadata-to-filter-mapper'
 
 import type { VuexContext } from '../flow.types'
 
@@ -15,7 +17,7 @@ const {attributes, sampleTable} = window.__INITIAL_STATE__ || {}
 export default {
   'GET_SUBJECT_METADATA' ({commit}: VuexContext) {
     api.get('/api/v2/' + sampleTable + '?includeCategories=true').then(response => {
-      const filterComponents = mappers.subjectMetadataToFilterMapper(response.meta)
+      const filterComponents = subjectMetadataToFilterMapper(response.meta)
       commit('SET_FILTER_COMPONENTS', filterComponents)
       commit('SET_TOTAL_NUMBER_OF_SAMPLES', response.total)
     })
@@ -26,7 +28,7 @@ export default {
 
     attributes.forEach(attribute => {
       promises.push(api.get('/api/v2/' + sampleTable + '?aggs=x==' + attribute.name).then(response => {
-        return mappers.aggregateDataToChartData(attribute, response.aggs)
+        return aggregateDataToChartDataMapper(attribute, response.aggs)
       }))
     })
 
@@ -39,9 +41,10 @@ export default {
     const promises = []
 
     attributes.forEach(attribute => {
-      const rsql = mappers.mapActiveFiltersToRSQL(state.activeFilters)
-      promises.push(api.get('/api/v2/' + sampleTable + '?aggs=x==' + attribute.name + '&q=' + rsql).then(response => {
-        return mappers.aggregateDataToChartData(attribute, response.aggs)
+      const filters = activeFiltersToRsqlMapper(state.activeFilters)
+      const rsql = filters === '' ? '' : '&q=' + filters
+      promises.push(api.get('/api/v2/' + sampleTable + '?aggs=x==' + attribute.name + rsql).then(response => {
+        return aggregateDataToChartDataMapper(attribute, response.aggs)
       }))
     })
