+++ src/utils/subject-metadata-to-filter-mapper.js
@@ -1,0 +1,71 @@
+const mapAttributesToFilters = (attribute) => {
+  switch (attribute.name) {
+    case 'biobank':
+      return {
+        id: attribute.name,
+        label: attribute.label,
+        filterType: 'dropdown',
+        options: attribute.categoricalOptions.map(option => ({
+          id: option.id,
+          label: option.label,
+          value: option.id
+        }))
+      }
+    case 'age_years':
+      return {
+        id: attribute.name,
+        label: attribute.label,
+        filterType: 'range',
+        options: {
+          min: 0,
+          max: 120
+        }
+      }
+    case 'sex':
+      return {
+        id: attribute.name,
+        label: attribute.label,
+        filterType: 'checkbox-row',
+        options: attribute.enumOptions.map(option => ({
+          id: option,
+          label: option,
+          value: option
+        }))
+      }
+    case 'smoking':
+      return {
+        id: attribute.name,
+        label: attribute.label,
+        filterType: 'checkbox-row',
+        options: [
+          {id: 'yes', label: 'Yes', value: true},
+          {id: 'no', label: 'No', value: false}
+        ]
+      }
+    case 'data_type':
+      return {
+        id: attribute.name,
+        label: attribute.label,
+        filterType: 'checkbox-list',
+        options: attribute.attributes.map(attribute => ({
+          id: attribute.name,
+          label: attribute.label,
+          value: attribute.name
+        }))
+      }
+  }
+}
+
+/**
+ * Create a map of sample category as key, and id, label, options object as value
+ */
+const subjectMetadataToFilterMapper = (metadata) => {
+  const filters = metadata.attributes.find(attribute => attribute.name === 'filters')
+  return filters.attributes.reduce((accumulator, attribute) => {
+    if (attribute.name === 'age') return accumulator // skip age column, we use age_years
+    accumulator[attribute.name] = mapAttributesToFilters(attribute)
+    return accumulator
+  }, {})
+}
+
+export default subjectMetadataToFilterMapper
